third_party/blink/web_tests/platform/generic/fast/css/border-width-large-expected.txt

Test for WebKit bug 18294 : Strange Result for getComputedStyle on borderWidth set in em

PASS computedStyle.getPropertyValue('border-width') is '12000px 11000px 10000px 9010px'
PASS computedStyle.borderTopWidth is '12000px'
PASS computedStyle.borderRightWidth is '11000px'
PASS computedStyle.borderBottomWidth is '10000px'
PASS computedStyle.borderLeftWidth is '9010px'
PASS successfullyParsed is true

TEST COMPLETE
